backend/pyproject.toml
[tool.poetry]
name = "frontend-challenge-backend"
version = "0.0.0"
description = ""
authors = ["nextml"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.101.1"
uvicorn = {version = "^0.21.1", extras = ["standard"]}
pillow = "^9.5.0"
orjson = "^3.6.0"
aiofiles = "^0.7.0"
httpx = "^0.24.0"
pydantic-settings = "^2.0.3"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.20" }
psycopg = "^3.1.8"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]
click = "^8.0.0"
autoflake = "^1.4"
pytest = "^6.2.3"
flake8 = "^3.9.2"
pre-commit = "^2.13.0"
ipykernel = "^5.5.5"
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
norecursedirs = ".venv __pycache__ .git .pytest_cache"
testpaths = [
    "app",
    "tests",
]
python_files = "*.py"
